refactor(controller): extract HPA reconciler to hpa_reconciler.go
Part 4/7 of splitting the god-controller into per-concern files.
Tracks #311.

Move the HorizontalPodAutoscaler lifecycle logic — creation, update,
removal, and Metal-accelerator skip — out of inferenceservice_controller.go
into a focused hpa_reconciler.go sibling.

Moved symbols:
  (*InferenceServiceReconciler).reconcileHPA
  (*InferenceServiceReconciler).constructHPA

Pure relocation — no signature changes, no behavior changes. Callers
in the same package continue to resolve the method references on the
receiver.

Line deltas (based on main; if earlier parts merge first the
controller file drops further on rebase):
  inferenceservice_controller.go: 1,567 → 1,412 (−155)
  hpa_reconciler.go:                      0 → 195 (new)

Verified: go vet clean, golangci-lint 0 issues, make test green,
internal/controller coverage unchanged at 83.1%.
